Implement content detail seasons and reviews endpoints

+++ src/main/java/com/selcan/StreamVibe/repository/ContentRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface ContentRepository extends JpaRepository<Content, Long> {
+public interface ContentRepository extends JpaRepository<Content, Integer> {
 
     List<Content> findByIsFeaturedTrue();
     Page<Content> findByTypeAndIsTrendingTrue(
